refactor: Full reorganization: Renamed and reorganized src files
