+++ src/debugger/variable/DebugVariable.ts
@@ -1,18 +1,12 @@
-import { DebugFrame } from './debugSessionTracker';
-import { EvalExpression } from './evalExpression';
-import { ImageVariable, ImageVariableType } from './imageVariable';
-import { VariableTypeFactory } from './variableTypeFactory';
+import { DebugFrame } from '../DebugFrame';
+import { ImageVariable } from './ImageVariable';
+import { ImageVariableType } from './ImageVariableType';
+import { BinaryInfo, DebugVariableType, DebugVariableArrayType, DebugVariableStructType } from './DebugVariableType';
+import { VariableTypeFactory } from './VariableTypeFactory';
+import { EvalExpression } from '../../utils/EvalExpression';
 
 export type DebugEntity = DebugVariable | ImageVariable;
 
-export interface IbinaryInfo<T> {
-    [key: string]: T,
-    sizeByte: T,
-    littleEndian: T,
-    signed: T,
-    fixedSize: T,
-    isInt: T,
-}
 export class DebugVariable {
     public category: string = "primitive";
     public isImageVariable: boolean = false;
@@ -25,14 +19,13 @@ export class DebugVariable {
     public endAddress: string | undefined;
     public sizeByte: string | undefined;
 
-    public binaryInfo: IbinaryInfo<any> = {
+    public binaryInfo: BinaryInfo = {
         sizeByte: 0,
         littleEndian: true,
         signed: true,
         fixedSize: false,
         isInt: true,
     };
-
 
     // value's type is the variable type or an array of DebugVariable or dictionary of DebugVariable
     public value: any | DebugVariableArrayType | DebugVariableStructType | undefined;
@@ -67,7 +60,6 @@ export class DebugVariable {
             //     this.endAddress = "0X" + (parseInt(this.meta.endAddress) + typeBinaryInfo.sizeByte).toString(16).toUpperCase();
             // }
         }
-
     }
 
     parse() {
@@ -127,22 +119,6 @@ export class DebugVariable {
         // if (this.meta === undefined) { return; }
         // if (until.depth === 0) { return; } // break if the depth reaches 0
 
-        // too much logs
-        if (0) {
-            console.log(
-                "drillDown",
-                this,
-                this.meta.name,
-                {
-                    variablesReference: this.meta.variablesReference,
-                    filter: undefined,
-                    start: 0,
-                    count: 1000
-                    // count: 0
-
-                });
-        }
-
         if (this.meta.variablesReference === 0) { return; }
 
         // check prevent drill down
@@ -164,7 +140,6 @@ export class DebugVariable {
         // fetch child debug variables
         let variables;
         try {
-
             variables = await this.frame.thread.tracker.session?.customRequest('variables',
                 {
                     variablesReference: this.meta.variablesReference,
@@ -303,51 +278,3 @@ export class DebugVariable {
         return ret;
     }
 }
-
-export class DebugVariableType {
-    public readonly name: string | undefined;
-    public readonly expression: string | undefined;
-    public isVisualizable: boolean = false;
-
-    private binaryMeta: IbinaryInfo<EvalExpression<any>>;
-
-    constructor(
-        name: string,
-        expression?: string,
-        binaryMetaString?: IbinaryInfo<string>,
-    ) {
-        this.name = name;
-        this.expression = expression;
-        this.binaryMeta = {
-            sizeByte: new EvalExpression<number>(binaryMetaString?.sizeByte || ""),
-            littleEndian: new EvalExpression<boolean>(binaryMetaString?.littleEndian || "true"),
-            signed: new EvalExpression<boolean>(binaryMetaString?.signed || "true"),
-            fixedSize: new EvalExpression<boolean>(binaryMetaString?.fixedSize || "false"),
-            isInt: new EvalExpression<boolean>(binaryMetaString?.isInt || "true"),
-        };
-    }
-
-    evalBinaryInfo(members: any): IbinaryInfo<any> {
-        const binaryInfo: IbinaryInfo<any> = {
-            sizeByte: 0,
-            littleEndian: false,
-            signed: false,
-            fixedSize: false,
-            isInt: true
-        };
-
-        Object.entries(this.binaryMeta).forEach(([key, evalExpression]) => {
-            binaryInfo[key] = evalExpression.eval(members);
-        });
-
-        return binaryInfo;
-    }
-}
-
-
-// Array type
-export type DebugVariableArrayType = DebugVariableType[];
-
-// Struct type
-export type DebugVariableStructType = { [key: string]: DebugVariableType };
-
