+++ src/debugger/variable/ImageVariable.ts
@@ -1,31 +1,19 @@
 import * as vscode from 'vscode';
-import { EvalExpression } from "./evalExpression";
-import { DebugVariable, DebugVariableType, IbinaryInfo } from "./debugVariable";
-import { VariableViewPanel } from '../panel';
 import sharp from 'sharp';
-// import * as cv from '@techstark/opencv-js';
 import * as fs from 'fs';
-import * as path from 'path';
 import * as crypto from 'crypto';
-import { DebugSessionTracker, DebugFrame } from './debugSessionTracker';
 
-export interface IimageInfo<T> {
-    [key: string]: T,
-    mem_width: T,
-    mem_height: T,
-    image_width: T,
-    image_height: T,
-    stride: T,
-    channels: T,
-    data: T,
-    format: T,
-    bytesForPx: T,
-}
+import { DebugSessionTracker } from '../DebugSessionTracker';
+import { DebugFrame } from '../DebugFrame';
+import { DebugVariable } from "./DebugVariable";
+import { DebugVariableType } from "./DebugVariableType";
+import { ImageVariableType, ImageInfo } from './ImageVariableType';
+import { EvalExpression } from "../../utils/EvalExpression";
 
 export class ImageVariable extends DebugVariable {
     public category: string = "image";
     public isImageVariable: boolean = true;
-    public imageInfo: IimageInfo<any> = {
+    public imageInfo: ImageInfo = {
         mem_width: 0,
         mem_height: 0,
         image_width: 0,
@@ -34,7 +22,7 @@ export class ImageVariable extends DebugVariable {
         channels: 1,
         data: "0X00",
         format: "RGB",
-        bytesForPx: "1",
+        bytesForPx: 1,
     };
 
     declare public type: ImageVariableType | undefined;
@@ -85,7 +73,6 @@ export class ImageVariable extends DebugVariable {
     async toFile() {
         console.log("begin toFile", this, this.name, this.expression);
         let buffer = this.buffer;
-
 
         this.updateImageInfo();
         this.updateBinaryInfo();
@@ -193,20 +180,6 @@ export class ImageVariable extends DebugVariable {
         console.log("imageArray:", imageArray);
 
 
-        // refer to https://github.com/Mohamed5341/opencv-image/blob/main/src/myimages.ts
-        // console.log("convert array to opencv image");
-        // let imageSrc = cv.matFromArray(this.imageInfo.mem_width, this.imageInfo.mem_height, this.type, imageArray);
-        // if (this.imageInfo.channels === 1) {
-        //     console.log("convert image to RGB from gray");
-        //     cv.cvtColor(imageSrc, imageSrc, cv.COLOR_GRAY2RGB);
-        // }
-
-        // if (this.imageInfo.channels === 3) {
-        //     cv.cvtColor(imageSrc, imageSrc, cv.COLOR_BGR2RGB);
-        // } else if (this.imageInfo.channels === 4) {
-        //     cv.cvtColor(imageSrc, imageSrc, cv.COLOR_BGRA2RGBA);
-        // }
-
         // Save
         const context = this.frame.thread.tracker.context;
         const storageUri = context.storageUri ? context.storageUri : context.globalStorageUri;
@@ -288,14 +261,6 @@ export class ImageVariable extends DebugVariable {
                 }
             });
 
-            // console.log("Setting opencv image to jimp");
-            // new Jimp({
-            //     width: this.imageInfo.mem_width, height: this.imageInfo.mem_height, data: Buffer.from(imageSrc.data)
-            // }
-            // ).write(filePath.fsPath);
-            // console.log("delete opencv image");
-            // imageSrc.delete();
-
             // Calc image hash
             // this.imageHash = crypto.createHash('blake2b512').update(bufferData).digest('hex');
             this.imageHash = crypto.createHash('md5').update(bufferData).digest('hex');
@@ -312,28 +277,6 @@ export class ImageVariable extends DebugVariable {
                 "imagewebUrl": "",
                 ...this.gatherMeta()
             };
-
-            // Display
-            // const openPath = vscode.Uri.file(filePath.toString()).toString().replace("/file:", "");
-            // vscode.commands.executeCommand('vscode.open', filePath.fsPath);
-            // console.log("rendering panel");
-            // VariableViewPanel.render(this.frame.thread.tracker.context);
-            // const panel = VariableViewPanel.currentPanel;
-            // if (panel) {
-            //     console.log("showing image on panel", panel);
-            //     const weburi = panel.getWebViewUrlString(filePath);
-            //     panel.postMessage({
-            //         command: "image",
-            //         // url: filePath.toString()
-            //         url: weburi,
-            //         meta: this.metaWide
-
-            //     });
-            //     panel.showPanel();
-            // }
-            // else {
-            //     console.log("panel is undefined");
-            // }
         }
 
         // Save .meta.json
@@ -353,62 +296,5 @@ export class ImageVariable extends DebugVariable {
             imageHiContPath: this.imageHiContPath,
             metaWide: this.metaWide,
         };
-    }
-
-
-}
-
-export class ImageVariableType extends DebugVariableType {
-    // ImageVariableType knows its member names or fixed values
-
-
-    private imageMeta: IimageInfo<EvalExpression<any>>;
-
-    constructor(
-        name: string,
-        expression?: string,
-        binaryMetaString?: IbinaryInfo<string>,
-        imageMetaString?: IimageInfo<string>,
-    ) {
-        super(name, expression, binaryMetaString);
-        this.imageMeta = {
-            mem_width: new EvalExpression<number>(imageMetaString?.mem_width || "0"),
-            mem_height: new EvalExpression<number>(imageMetaString?.mem_height || "0"),
-            image_width: new EvalExpression<number>(imageMetaString?.image_width || "0"),
-            image_height: new EvalExpression<number>(imageMetaString?.image_height || "0"),
-            stride: new EvalExpression<number>(imageMetaString?.stride || "0"),
-            channels: new EvalExpression<number>(imageMetaString?.channels || "3"),
-            data: new EvalExpression<string>(imageMetaString?.data || "'0X00'"),
-            format: new EvalExpression<string>(imageMetaString?.format || "'RGB'"),
-            bytesForPx: new EvalExpression<number>(imageMetaString?.bytesForPx || "1"),
-        };
-    }
-
-    evalImageInfo(members: any): IimageInfo<any> {
-        const imageInfo: IimageInfo<any> = {
-            mem_width: 0,
-            mem_height: 0,
-            image_width: 0,
-            image_height: 0,
-            stride: 0,
-            channels: 0,
-            data: "'0X00'",
-            format: "'RGB'",
-            bytesForPx: 1,
-        };
-
-        Object.entries(this.imageMeta).forEach(([key, evalExpression]) => {
-            imageInfo[key] = evalExpression.eval(members);
-        });
-
-        imageInfo.mem_width = parseInt(imageInfo.mem_width);
-        imageInfo.mem_height = parseInt(imageInfo.mem_height);
-        imageInfo.image_width = parseInt(imageInfo.image_width);
-        imageInfo.image_height = parseInt(imageInfo.image_height);
-        imageInfo.stride = parseInt(imageInfo.stride);
-        imageInfo.channels = parseInt(imageInfo.channels);
-        imageInfo.bytesForPx = parseInt(imageInfo.bytesForPx);
-
-        return imageInfo;
     }
 }
