+++ src/panel/VariableViewPanel.ts
