+++ src/tracker/VariableTrackerRegister.ts
@@ -1,0 +1,28 @@
+import * as vscode from 'vscode';
+
+import { VariableTracker } from './VariableTracker';
+import { DebugSessionTracker } from '../debugger/DebugSessionTracker';
+
+export class VariableTrackerRegister implements vscode.DebugAdapterTrackerFactory {
+    private variableTracker: VariableTracker;
+    public readonly context: vscode.ExtensionContext;
+
+    constructor(_context: vscode.ExtensionContext) {
+        this.variableTracker = new VariableTracker(_context);
+        this.context = _context;
+    }
+
+    createDebugAdapterTracker(session: vscode.DebugSession):
+        vscode.ProviderResult<vscode.DebugAdapterTracker> {
+        console.log("sessoin!", session);
+        const tracker = DebugSessionTracker.newSessionTracker(this.context, session);
+        DebugSessionTracker.breakCount = 0; // FIXME: It should better to manage not by a static.
+        this.variableTracker.breakpoints = [];
+        return this.variableTracker;
+    }
+
+    static register(context: vscode.ExtensionContext): vscode.Disposable {
+        console.log(context);
+        return vscode.debug.registerDebugAdapterTrackerFactory('*', new VariableTrackerRegister(context));
+    }
+}
